move tests to test_io_hdf5tools add test for Data.set_data_io
